+++ config/config.toml
@@ -140,6 +140,14 @@ kafkaDialTimeoutInSec = 10
 retryTimeWindowInMins = 180
 minRetryBackoffInS = 10
 maxRetryBackoffInS = 300
+    # settings similar to below one can be used to throttle router
+    # sending events to downstream destinations
+    # [Router.throttler.AM]
+    # eventLimit = 1000
+    # timeWindowInS =          1
+    # userLevelThrottling =    true
+    # userLevelLimit =         10
+    # userLevelTimeWindowInS = 1
 
 [BatchRouter]
 mainLoopSleepInS = 2
